+++ admin/src/scenes/missions/view/index.js
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from "react";
-import { Switch, Route, useHistory } from "react-router-dom";
+import { Switch, useHistory } from "react-router-dom";
+import { SentryRoute } from "../../../sentry";
 
 import api from "../../../services/api";
 import Details from "./details";
@@ -67,10 +68,10 @@ export default function Index({ ...props }) {
     <>
       <Breadcrumbs items={[{ label: "Missions", to: "/mission" }, { label: "Fiche de la mission" }]} />
       <Switch>
-        <Route path="/mission/:id/youngs" component={() => <Youngs mission={mission} applications={applications} updateApplications={fetchApplication} />} />
-        <Route path="/mission/:id/historique" component={() => <Historic mission={mission} />} />
-        <Route path="/mission/:id/propose-mission" component={() => <ProposeMission mission={mission} />} />
-        <Route path="/mission/:id" component={() => <Details mission={mission} structure={structure} tutor={tutor} />} />
+        <SentryRoute path="/mission/:id/youngs" component={() => <Youngs mission={mission} applications={applications} updateApplications={fetchApplication} />} />
+        <SentryRoute path="/mission/:id/historique" component={() => <Historic mission={mission} />} />
+        <SentryRoute path="/mission/:id/propose-mission" component={() => <ProposeMission mission={mission} />} />
+        <SentryRoute path="/mission/:id" component={() => <Details mission={mission} structure={structure} tutor={tutor} />} />
       </Switch>
     </>
   );